+++ src/extensions/base/cmbar.cpp
@@ -19,11 +19,11 @@ public:
     populateColorMap<CmBarMode::Colors>(cMap);
   }
 
-  int indent(Buffer& buf, int line) { return 0; }
+  size_t indent(Buffer& buf, size_t line) { return 0; }
 
   KeyCmdMap& getKeyCmdMap() { return kcMap; }
   ColorMap& getColorMap() { return cMap; }
-  void getColorFor(AttrColor& fg, AttrColor& bg, int lineNum, int pos,
+  void getColorFor(AttrColor& fg, AttrColor& bg, size_t lineNum, size_t pos,
                    const Buffer& b, bool isHighlighted) {
     fg = cMap.get(isHighlighted ? "highlightfg" : "defaultfg");
     bg = cMap.get(isHighlighted ? "highlightbg" : "defaultbg");
