+++ src/extensions/cpp/mode.cpp
@@ -13,21 +13,21 @@ public:
     populateColorMap<CppMode::Colors>(getColorMap());
   }
 
-  int indent(Buffer& buf, int line) {
+  size_t indent(Buffer& buf, size_t line) {
     if(!(0 < line && line < buf.length())) return 0;
     const auto& prev = buf.at(line-1);
     auto& curr = buf.at(line);
     const auto& prevLine = prev.get();
     const auto& currLine = curr.get();
     Pcre nspace("namespace .*?{");
-    int prevInd;
+    size_t prevInd;
     if(nspace.isMatch(prevLine))
       prevInd = 0;
     else if(prevLine[0] == '#')
       prevInd = 0;
     else
       prevInd = prev.indentSize();
-    int currInd = curr.indentSize();
+    size_t currInd = curr.indentSize();
     if(currLine[0] == '#') return -currInd;
     return prevInd - currInd;
   }
